requirements.txt
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
anthropic>=0.40.0
pydantic>=2.0
python-dotenv>=1.0
jinja2>=3.1
requests>=2.31
